+++ src/components/atoms/icons/icons.ts
@@ -1,0 +1,19 @@
+import BanIcon from '@/components/atoms/icons/library/BanIcon.vue';
+import ChevronIcon from '@/components/atoms/icons/library/ChevronIcon.vue';
+import ClockIcon from '@/components/atoms/icons/library/ClockIcon.vue';
+import ConfettiIcon from '@/components/atoms/icons/library/ConfettiIcon.vue';
+import InboxIcon from '@/components/atoms/icons/library/InboxIcon.vue';
+import SearchIcon from '@/components/atoms/icons/library/SearchIcon.vue';
+import StarIcon from '@/components/atoms/icons/library/StarIcon.vue';
+import UserIcon from '@/components/atoms/icons/library/UserIcon.vue';
+
+export const icons = {
+  BanIcon,
+  ChevronIcon,
+  ClockIcon,
+  ConfettiIcon,
+  InboxIcon,
+  SearchIcon,
+  StarIcon,
+  UserIcon
+};