+++ src/components/patterns/main-header/MainHeader.vue
@@ -1,8 +1,8 @@
 <template>
   <HeaderBase class="h-12 pb-2 flex justify-between items-center">
-    <TitleSection>
+    <TitleMainSection>
       <slot></slot>
-    </TitleSection>
+    </TitleMainSection>
 
     <UserAvatar class="h-8 w-8">
       <img src="@/assets/profile-avatar.png" alt="User avatar" />
@@ -12,6 +12,6 @@
 
 <script setup lang="ts">
 import HeaderBase from '@/components/atoms/header/HeaderBase.vue';
-import TitleSection from '@/components/patterns/section/TitleSection.vue';
+import TitleMainSection from '@/components/patterns/main-section/TitleMainSection.vue';
 import UserAvatar from '@/components/patterns/user-avatar/UserAvatar.vue';
 </script>