+++ src/views/OnBoardingView.vue
@@ -1,0 +1,11 @@
+<template>
+  <TemplateBase>
+    <template #title>OnBoarding</template>
+
+    <template #content></template>
+  </TemplateBase>
+</template>
+
+<script setup lang="ts">
+import TemplateBase from '@/components/templates/TemplateBase.vue';
+</script>